docs: proper key to get indicator name
In line 26 when using indicator.name no results appear but using indicator.displayName fixes the issue since in the json file the key is "displayName" not "name" and in the https://runtime.dhis2.nu/#/hooks/useDataQuery example too! I don't know if this was on purpose since it wasn't mentioned in the instructions. Thanks!